Auto lightup classification/bracematching on likely strings
