+++ src/EditorFeatures/Core/EmbeddedLanguages/Json/JsonEmbeddedBraceMatcher.cs
@@ -39,7 +39,7 @@ namespace Microsoft.CodeAnalysis.Editor.EmbeddedLanguages.Json
             var semanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
 
             var detector = JsonLanguageDetector.GetOrCreate(semanticModel.Compilation, _info);
-            var tree = detector.TryParseString(token, semanticModel, cancellationToken);
+            var tree = detector.TryParseString(token, semanticModel, includeProbableStrings: true, cancellationToken);
             if (tree == null)
                 return null;
