+++ src/Features/Core/Portable/EmbeddedLanguages/Json/LanguageServices/AbstractJsonDetectionAnalyzer.cs
@@ -75,9 +75,13 @@ namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.Json.LanguageService
                 else
                 {
                     var token = child.AsToken();
+
+                    // If we have a string, and it's not being passed to a known JSON api (and it doesn't have a
+                    // lang=json comment), but it is parseable as JSON with enough structure that we are confident it is
+                    // json, then report that json features could light up here.
                     if (_info.IsAnyStringLiteral(token.RawKind) &&
-                        detector.TryParseString(token, context.SemanticModel, cancellationToken) == null &&
-                        IsProbablyJson(token))
+                        detector.TryParseString(token, context.SemanticModel, includeProbableStrings: false, cancellationToken) == null &&
+                        detector.IsProbablyJson(token, out _))
                     {
                         var chars = _info.VirtualCharService.TryConvertToVirtualChars(token);
                         var strictTree = JsonParser.TryParse(chars, JsonOptions.Strict);
@@ -96,37 +100,6 @@ namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.Json.LanguageService
                     }
                 }
             }
-        }
-
-        public bool IsProbablyJson(SyntaxToken token)
-        {
-            var chars = _info.VirtualCharService.TryConvertToVirtualChars(token);
-            var tree = JsonParser.TryParse(chars, JsonOptions.Loose);
-            if (tree == null || !tree.Diagnostics.IsEmpty)
-                return false;
-
-            return ContainsProbableJsonObject(tree.Root);
-        }
-
-        private static bool ContainsProbableJsonObject(JsonNode node)
-        {
-            if (node.Kind == JsonKind.Object)
-            {
-                var objNode = (JsonObjectNode)node;
-                if (objNode.Sequence.Length >= 1)
-                    return true;
-            }
-
-            foreach (var child in node)
-            {
-                if (child.IsNode)
-                {
-                    if (ContainsProbableJsonObject(child.Node))
-                        return true;
-                }
-            }
-
-            return false;
         }
     }
 }
