+++ src/Features/Core/Portable/EmbeddedLanguages/Json/LanguageServices/AbstractJsonDiagnosticAnalyzer.cs
@@ -71,7 +71,7 @@ namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.Json.LanguageService
                     var token = child.AsToken();
                     if (_info.IsAnyStringLiteral(token.RawKind))
                     {
-                        var tree = detector.TryParseString(token, context.SemanticModel, cancellationToken);
+                        var tree = detector.TryParseString(token, context.SemanticModel, includeProbableStrings: false, cancellationToken);
                         if (tree != null)
                         {
                             foreach (var diag in tree.Diagnostics)
