+++ src/Features/Core/Portable/EmbeddedLanguages/Json/LanguageServices/JsonLanguageDetector.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text.Json;
@@ -76,6 +77,62 @@ namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.Json.LanguageService
 
         protected override JsonTree? TryParse(VirtualCharSequence chars, JsonOptions options)
             => JsonParser.TryParse(chars, options);
+
+        /// <inheritdoc cref="TryParseString(SyntaxToken, SemanticModel, bool, CancellationToken)"/>
+        /// <summary>
+        /// If <paramref name="includeProbableStrings"/> is true, then this will also succeed on a string-literal like
+        /// <paramref name="token"/> that strongly appears to have JSON in it.  This allows some features to light up
+        /// automatically on code that is strongly believed to be JSON, but which is not passed to a known JSON api,
+        /// and does not have a comment on it stating it is JSON.
+        /// </summary>
+        public JsonTree? TryParseString(SyntaxToken token, SemanticModel semanticModel, bool includeProbableStrings, CancellationToken cancellationToken)
+        {
+            var result = TryParseString(token, semanticModel, cancellationToken);
+            if (result != null)
+                return result;
+
+            if (includeProbableStrings && IsProbablyJson(token, out var tree))
+                return tree;
+
+            return null;
+        }
+
+        /// <summary>
+        /// Returns <see langword="true"/> if this string-like <paramref name="token"/> is likely a JSON literal.  As
+        /// many simple strings are legal JSON (like <c>0</c>) we require enough structure here to feel confident that
+        /// this truly is JSON.  Currently, this means it must have at least one <c>{ ... }</c> object literal, and that
+        /// literal must have at least one <c>"prop": val</c> property in it.
+        /// </summary>
+        public bool IsProbablyJson(SyntaxToken token, [NotNullWhen(true)] out JsonTree? tree)
+        {
+            var chars = this.Info.VirtualCharService.TryConvertToVirtualChars(token);
+            tree = JsonParser.TryParse(chars, JsonOptions.Loose);
+            if (tree == null || !tree.Diagnostics.IsEmpty)
+                return false;
+
+            return ContainsProbableJsonObject(tree.Root);
+        }
+
+        private static bool ContainsProbableJsonObject(JsonNode node)
+        {
+            if (node.Kind == JsonKind.Object)
+            {
+                var objNode = (JsonObjectNode)node;
+                if (objNode.Sequence.Length >= 1)
+                    return true;
+            }
+
+            foreach (var child in node)
+            {
+                if (child.IsNode)
+                {
+                    if (ContainsProbableJsonObject(child.Node))
+                        return true;
+                }
+            }
+
+            return false;
+        }
 
         protected override bool IsArgumentToWellKnownAPI(
             SyntaxToken token,
